feat: factor the api code into files for each endpoint
